+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.1] - 2024-08-08
+
+- Full rewrite using dependency injection a la @lanayx
+- GTK4 GUI (using gir.core)
+- Midjourney support using https://imaginepro.ai API
+- Command palette with simplified navigation
+- Introduction message
+
 ## [0.0.8] - 2024-07-19
 
 Added: